docs: updated README.md with app architecture and dependencies, refactored router package
